requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
transformers
pgvector
sqlalchemy
psycopg2-binary
python-dotenv
faiss-cpu
PyPDF2
python-multipart
asyncpg
langchain-community
rank_bm25
spacy
apscheduler
black
pytest
pytest-asyncio
pytest-cov
email-validator
gunicorn
passlib
PyJWT